requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.5
django-debug-toolbar==4.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
djangorestframework==3.14.0
idna==3.6
Pillow==10.0.1
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
